[rs-order] feat: Add quota count to basket

+++ rs-order/order/order-domain/src/main/java/fr/cnes/regards/modules/order/domain/DatasetTask.java
@@ -76,7 +76,7 @@ public class DatasetTask extends AbstractReliantTask<FilesTask> implements Compa
     private int objectsCount = 0;
 
     @Column(name = "files_count")
-    private int filesCount = 0;
+    private long filesCount = 0;
 
     @Column(name = "files_size")
     private long filesSize = 0;
@@ -111,11 +111,11 @@ public class DatasetTask extends AbstractReliantTask<FilesTask> implements Compa
         this.objectsCount = objectsCount;
     }
 
-    public int getFilesCount() {
+    public long getFilesCount() {
         return filesCount;
     }
 
-    public void setFilesCount(int filesCount) {
+    public void setFilesCount(long filesCount) {
         this.filesCount = filesCount;
     }
